move loadStories into scripts/ - just to help organize
