+++ views/partials/social_media_scripts.ejs
@@ -1,3 +1,4 @@
+<!-- THE SCRIPTS IN THIS PARTIAL CORRESPOND TO THE HTML ELEMENTS IN ./social_media.ejs -->
 <!-- FACEBOOK -->
 <script>(function(d, s, id) {
   var js, fjs = d.getElementsByTagName(s)[0];
